restructure position, give each component a seperate directory
