+++ ClientApp/src/app/position/position.component.ts
@@ -1,12 +1,10 @@
 import { Component, OnInit, Inject } from '@angular/core';
-import { PositionApiService, Position, PositionEdit, PersonAssignment, PositionHolder } from '../services/positionapi.service';
+import { PositionApiService, Position, PositionEdit, PositionHolder } from '../services/positionapi.service';
 import { Router } from '@angular/router';
 import { MatDialogRef, MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
-import { FormBuilder, Validators, AbstractControl } from '@angular/forms';
-import { Observable, of } from 'rxjs';
-import { startWith, switchMap, map } from 'rxjs/operators';
-import * as moment from "moment";
-import { AssignmentOption } from '../common/search-select/search-select.component';
+import { PositionAssignDialogComponent } from './position-assign-dialog/position-assign-dialog.component';
+import { PositionDismissDialogComponent } from "./position-dismiss-dialog/position-dismiss-dialog.component";
+import { PositionEditDialogComponent } from "./position-edit-dialog/position-edit-dialog.component";
 
 @Component({
   selector: 'app-position',
@@ -52,7 +50,7 @@ export class PositionComponent implements OnInit {
       shortName: position.shortName,
       positionID: position.positionID,
     };
-    const dialogRef = this.dialog.open(PositionEditCialogComponent, {
+    const dialogRef = this.dialog.open(PositionEditDialogComponent, {
       width: '400px',
       data: posEdit,
     });
@@ -82,205 +80,6 @@ export class PositionComponent implements OnInit {
         position,
         person,
       },
-    });
-  }
-}
-
-@Component({
-  selector: 'app-position-edit-dialog',
-  templateUrl: 'position-edit-dialog.html',
-})
-export class PositionEditCialogComponent {
-
-  // track state
-  public savingBeforeClose = false;
-
-  constructor(
-    private positionApiService: PositionApiService,
-    public dialogRef: MatDialogRef<PositionEditCialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: PositionEdit) {
-      this.savingBeforeClose = false;
-    }
-
-  onNoClick(): void {
-    this.dialogRef.close();
-  }
-
-  onSave(): void {
-    if (this.savingBeforeClose) { return; }
-    console.log('saving...');
-    this.savingBeforeClose = true;
-    this.dialogRef.disableClose = true;
-    this.positionApiService.update(this.data).subscribe(val => {
-      this.dialogRef.close(this.data);
-    }, err => {
-      // how do we want to handle errors? Notification top right?
-      console.log(err);
-      this.dialogRef.close();
-    });
-  }
-}
-
-@Component({
-  selector: 'app-position-create-dialog',
-  templateUrl: 'position-create-dialog.html',
-})
-export class PositionCreateDialogComponent {
-
-  // track state
-  public savingBeforeClose = false;
-  public data: PositionEdit;
-
-  constructor(
-    private positionApiService: PositionApiService,
-    public dialogRef: MatDialogRef<PositionEditCialogComponent>) {
-      this.savingBeforeClose = false;
-      this.data = {
-        name: '',
-        shortName: '',
-        positionID: -1, // ignored by api anyways
-      };
-    }
-
-  onNoClick(): void {
-    this.dialogRef.close();
-  }
-
-  onSave(): void {
-    if (this.savingBeforeClose) { return; }
-    console.log('saving...');
-    this.savingBeforeClose = true;
-    this.dialogRef.disableClose = true;
-    this.positionApiService.create(this.data).subscribe(val => {
-      this.dialogRef.close(this.data);
-    }, err => {
-      // how do we want to handle errors? Notification top right?
-      console.log(err);
-      this.dialogRef.close();
-    });
-  }
-}
-
-@Component({
-  selector: 'app-position-assign-dialog',
-  templateUrl: 'position-assign-dialog.html',
-})
-export class PositionAssignDialogComponent implements OnInit {
-
-  // track state
-  public savingBeforeClose = false;
-  public assignSuggestions: AssignmentOption[] = [];
-
-  constructor(
-    private formBuilder: FormBuilder,
-    private positionApiService: PositionApiService,
-    public dialogRef: MatDialogRef<PositionEditCialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public position: Position) {
-      this.savingBeforeClose = false;
-  }
-
-  ngOnInit(): void {
-    this.positionApiService.getAssignmentsSuggestion().subscribe(suggestions => {
-      this.assignSuggestions = suggestions.map(s => {
-        return {name: s.name, id: s.personID};
-      });
-    });
-    this.assignDate.setValue(moment());
-    this.assignForm.valueChanges.subscribe(v => {
-      console.log(v);
-    })
-  }
-
-  onNoClick(): void {
-    this.dialogRef.close();
-  }
-
-  assignForm = this.formBuilder.group({
-    assignPerson: [true, Validators.required],
-    assignDate: [null, Validators.required],
-  });
-
-  get assignPerson() {
-    return this.assignForm.get('assignPerson');
-  }
-
-  get assignDate() {
-    return this.assignForm.get('assignDate');
-  }
-
-  onSubmit(): void {
-    if (this.savingBeforeClose) { return; }
-    // triggers errors
-    this.assignForm.markAllAsTouched();
-    if (this.assignForm.invalid) { return; }
-    console.log('saving...');
-    this.savingBeforeClose = true;
-    this.dialogRef.disableClose = true;
-
-    const personsToAssign: number[] = [];
-    const personID = this.assignPerson.value.id;
-    personsToAssign.push(personID);
-    this.positionApiService.assign(this.position.positionID, this.assignDate.value.toJSON(), personsToAssign).subscribe(val => {
-      this.dialogRef.close();
-    }, err => {
-      // how do we want to handle errors? Notification top right?
-      console.log(err);
-      this.dialogRef.close();
-    });
-  }
-}
-
-@Component({
-  selector: 'app-position-dismiss-dialog',
-  templateUrl: 'position-dismiss-dialog.html',
-})
-export class PositionDismissDialogComponent implements OnInit{
-  // track state
-  public savingBeforeClose = false;
-
-  constructor(
-    private formBuilder: FormBuilder,
-    private positionApiService: PositionApiService,
-    public dialogRef: MatDialogRef<PositionDismissDialogComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: {position: Position, person: PositionHolder}) {
-      this.savingBeforeClose = false;
-  }
-
-  ngOnInit(): void {
-    this.date.setValue(moment());
-    this.date.valueChanges.subscribe(v => {
-      console.log(v);
-    });
-  }
-
-  onNoClick(): void {
-    this.dialogRef.close();
-  }
-
-  dismissForm = this.formBuilder.group({
-    date: [null, Validators.required],
-  });
-
-  get date() {
-    return this.dismissForm.get('date');
-  }
-
-  onSubmit(): void {
-    if (this.savingBeforeClose) { return; }
-    // triggers errors
-    this.dismissForm.markAllAsTouched();
-    if (this.dismissForm.invalid) { return; }
-    console.log('saving...');
-    this.savingBeforeClose = true;
-    this.dialogRef.disableClose = true;
-
-    const personsToDismiss: number[] = [this.data.person.personID];
-    this.positionApiService.dismiss(this.data.position.positionID, this.date.value.toJSON(), personsToDismiss).subscribe(val => {
-      this.dialogRef.close();
-    }, err => {
-      // how do we want to handle errors? Notification top right?
-      console.log(err);
-      this.dialogRef.close();
     });
   }
 }
